+++ package.json
@@ -66,6 +66,7 @@
     ]
   },
   "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@types/express": "^4.17.20",
     "@types/jest": "^29.5.8",
     "@types/node": "^20.8.7",
